+++ server/marketplace/Model/MarketplacePurchase.cs
@@ -9,6 +9,6 @@ public class MarketplacePurchase : BaseModel
     public string ListingId { get; set; }
     public uint BuyerId { get; set; }
     public ushort PurchaseCount { get; set; }
-    public uint PurchasePrice { get; set; }
+    public ulong PurchasePrice { get; set; }
     public new DateTime CreatedDate { get; set; }
 }
